refactor(architecture): remove infra service package
