+++ internal/app/openbaocluster/deletionops/cleanup.go
@@ -1,0 +1,77 @@
+package deletionops
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/go-logr/logr"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	openbaov1alpha1 "github.com/dc-tec/openbao-operator/api/v1alpha1"
+	"github.com/dc-tec/openbao-operator/internal/platform/constants"
+	portopenbao "github.com/dc-tec/openbao-operator/internal/port/openbao"
+)
+
+// Cleanup handles deletion side effects that are not covered by OwnerReferences.
+func Cleanup(
+	ctx context.Context,
+	logger logr.Logger,
+	kubeClient client.Client,
+	cluster *openbaov1alpha1.OpenBaoCluster,
+	policy openbaov1alpha1.DeletionPolicy,
+) error {
+	if policy == "" {
+		policy = openbaov1alpha1.DeletionPolicyRetain
+	}
+
+	logger = logger.WithValues("deletionPolicy", string(policy))
+	logger.Info(
+		"Processing cleanup for deleted OpenBaoCluster",
+		"note",
+		"Most resources are deleted by Kubernetes GC via OwnerReferences",
+	)
+
+	if policy == openbaov1alpha1.DeletionPolicyDeletePVCs || policy == openbaov1alpha1.DeletionPolicyDeleteAll {
+		if err := deletePVCs(ctx, kubeClient, cluster); err != nil {
+			return fmt.Errorf("failed to delete PVCs for OpenBaoCluster %s/%s: %w", cluster.Namespace, cluster.Name, err)
+		}
+		logger.Info("PVCs deleted per deletion policy")
+	} else {
+		logger.Info("Preserving PVCs per Retain policy")
+	}
+
+	return nil
+}
+
+func deletePVCs(ctx context.Context, kubeClient client.Client, cluster *openbaov1alpha1.OpenBaoCluster) error {
+	var pvcList corev1.PersistentVolumeClaimList
+	if err := kubeClient.List(
+		ctx,
+		&pvcList,
+		client.InNamespace(cluster.Namespace),
+		client.MatchingLabels(map[string]string{constants.LabelOpenBaoCluster: cluster.Name}),
+	); err != nil {
+		return err
+	}
+
+	for i := range pvcList.Items {
+		pvc := &pvcList.Items[i]
+		if portopenbao.UsesExistingACMESharedCache(cluster) && pvc.Name == portopenbao.ACMESharedCacheClaimName(cluster) {
+			continue
+		}
+		if len(pvc.Finalizers) > 0 {
+			original := pvc.DeepCopy()
+			pvc.Finalizers = nil
+			if err := kubeClient.Patch(ctx, pvc, client.MergeFrom(original)); err != nil && !apierrors.IsNotFound(err) {
+				return err
+			}
+		}
+		if err := kubeClient.Delete(ctx, pvc); err != nil && !apierrors.IsNotFound(err) {
+			return err
+		}
+	}
+
+	return nil
+}
