--- internal/service/infra/compute_test.go
@@ -1,872 +1,0 @@
-//go:build integration
-// +build integration
-
-package infra
-
-import (
-	"context"
-	"fmt"
-	"strings"
-	"testing"
-
-	"github.com/dc-tec/openbao-operator/internal/platform/constants"
-	portopenbao "github.com/dc-tec/openbao-operator/internal/port/openbao"
-	"github.com/go-logr/logr"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-
-	openbaov1alpha1 "github.com/dc-tec/openbao-operator/api/v1alpha1"
-)
-
-const (
-	testOpenBaoLivenessProbeTimeout  = "4s"
-	testOpenBaoReadinessProbeTimeout = "10s"
-	testOpenBaoStartupProbeTimeout   = "5s"
-	testOpenBaoProbeAddr             = "https://127.0.0.1:8200"
-)
-
-func TestStatefulSetStartsWithOneReplicaWhenNotInitialized(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-init", ns)
-	cluster.Status.Initialized = false
-	cluster.Spec.Replicas = 3
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if statefulSet.Spec.Replicas == nil {
-		t.Fatalf("expected StatefulSet to have replicas set")
-	}
-
-	// Should start with 1 replica when not initialized
-	if *statefulSet.Spec.Replicas != 1 {
-		t.Fatalf("expected StatefulSet replicas to be 1 when not initialized, got %d", *statefulSet.Spec.Replicas)
-	}
-}
-
-func TestStatefulSetScalesToDesiredReplicasWhenInitialized(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-scaled", ns)
-	cluster.Status.Initialized = true
-	cluster.Spec.Replicas = 3
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if statefulSet.Spec.Replicas == nil {
-		t.Fatalf("expected StatefulSet to have replicas set")
-	}
-
-	// Should scale to desired replicas when initialized
-	if *statefulSet.Spec.Replicas != cluster.Spec.Replicas {
-		t.Fatalf("expected StatefulSet replicas to be %d when initialized, got %d", cluster.Spec.Replicas, *statefulSet.Spec.Replicas)
-	}
-}
-
-func TestStatefulSetReplicaScalingTableDriven(t *testing.T) {
-	tests := []struct {
-		name         string
-		initialized  bool
-		specReplicas int32
-		wantReplicas int32
-	}{
-		{
-			name:         "not initialized starts with 1 replica",
-			initialized:  false,
-			specReplicas: 3,
-			wantReplicas: 1,
-		},
-		{
-			name:         "initialized scales to desired replicas",
-			initialized:  true,
-			specReplicas: 3,
-			wantReplicas: 3,
-		},
-		{
-			name:         "initialized with 1 replica stays at 1",
-			initialized:  true,
-			specReplicas: 1,
-			wantReplicas: 1,
-		},
-		{
-			name:         "not initialized with 1 replica stays at 1",
-			initialized:  false,
-			specReplicas: 1,
-			wantReplicas: 1,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			k8sClient, _ := envtestClientForPackage(t)
-			manager := NewManager(k8sClient)
-
-			ns := testNamespace(t)
-			cluster := newMinimalCluster("test-replica", ns)
-			cluster.Spec.Replicas = tt.specReplicas
-			cluster.Status.Initialized = tt.initialized
-			createClusterCRForTest(t, k8sClient, cluster)
-
-			// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-			createTLSSecretForTest(t, k8sClient, cluster)
-
-			ctx := context.Background()
-
-			spec := newTestStatefulSetSpec(cluster)
-			if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-				t.Fatalf("Reconcile() error = %v", err)
-			}
-
-			sts := &appsv1.StatefulSet{}
-			err := k8sClient.Get(ctx, types.NamespacedName{
-				Namespace: cluster.Namespace,
-				Name:      statefulSetName(cluster),
-			}, sts)
-			if err != nil {
-				t.Fatalf("expected StatefulSet to exist: %v", err)
-			}
-
-			if sts.Spec.Replicas == nil {
-				t.Fatalf("expected StatefulSet to have replicas set")
-			}
-
-			if *sts.Spec.Replicas != tt.wantReplicas {
-				t.Fatalf("StatefulSet replicas = %d, want %d", *sts.Spec.Replicas, tt.wantReplicas)
-			}
-		})
-	}
-}
-
-func TestStatefulSetDoesNotUpdateVolumeClaimTemplatesOnStorageResize(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-storage-resize", ns)
-	cluster.Status.Initialized = true
-	cluster.Spec.Storage.Size = "10Gi"
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: cluster.Namespace, Name: statefulSetName(cluster)}, statefulSet); err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if len(statefulSet.Spec.VolumeClaimTemplates) != 1 {
-		t.Fatalf("expected one VolumeClaimTemplate, got %d", len(statefulSet.Spec.VolumeClaimTemplates))
-	}
-	gotBefore := statefulSet.Spec.VolumeClaimTemplates[0].Spec.Resources.Requests[corev1.ResourceStorage]
-	if gotBefore.Cmp(resource.MustParse("10Gi")) != 0 {
-		t.Fatalf("expected template size 10Gi before resize, got %s", gotBefore.String())
-	}
-
-	// Update desired storage size in the CR.
-	cluster.Spec.Storage.Size = "20Gi"
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() after resize error = %v", err)
-	}
-
-	statefulSetAfter := &appsv1.StatefulSet{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: cluster.Namespace, Name: statefulSetName(cluster)}, statefulSetAfter); err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if len(statefulSetAfter.Spec.VolumeClaimTemplates) != 1 {
-		t.Fatalf("expected one VolumeClaimTemplate, got %d", len(statefulSetAfter.Spec.VolumeClaimTemplates))
-	}
-	gotAfter := statefulSetAfter.Spec.VolumeClaimTemplates[0].Spec.Resources.Requests[corev1.ResourceStorage]
-	if gotAfter.Cmp(resource.MustParse("10Gi")) != 0 {
-		t.Fatalf("expected template size to remain 10Gi after resize, got %s", gotAfter.String())
-	}
-}
-
-//nolint:gocyclo // This is a comprehensive, table-free assertion test for container configuration.
-func TestStatefulSetHasCorrectContainerConfiguration(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-container", ns)
-	uiEnabled := true
-	cluster.Spec.Configuration = &openbaov1alpha1.OpenBaoConfiguration{
-		UI: &uiEnabled,
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if len(statefulSet.Spec.VolumeClaimTemplates) != 1 {
-		t.Fatalf("expected one VolumeClaimTemplate, got %d", len(statefulSet.Spec.VolumeClaimTemplates))
-	}
-
-	containerFound := false
-	for _, c := range statefulSet.Spec.Template.Spec.Containers {
-		if c.Name == constants.ContainerBao {
-			containerFound = true
-			if c.Image != cluster.Spec.Image {
-				t.Fatalf("expected container image %q, got %q", cluster.Spec.Image, c.Image)
-			}
-			if len(c.VolumeMounts) == 0 {
-				t.Fatalf("expected container to have volume mounts")
-			}
-
-			// Verify container runs via the wrapper and passes bao server with rendered config path.
-			if len(c.Command) != 1 || c.Command[0] != "/utils/bao-wrapper" {
-				t.Fatalf("expected container command to be /utils/bao-wrapper, got %v", c.Command)
-			}
-
-			foundChildCmd := false
-			for i := range c.Args {
-				if c.Args[i] == "--" && len(c.Args) > i+3 &&
-					c.Args[i+1] == openBaoBinaryName &&
-					c.Args[i+2] == "server" &&
-					strings.Contains(c.Args[i+3], "-config=") {
-					foundChildCmd = true
-					break
-				}
-			}
-			if !foundChildCmd {
-				t.Fatalf("expected wrapper args to include %s server -config=..., got %v", openBaoBinaryName, c.Args)
-			}
-
-			foundSATokenMount := false
-			for _, mount := range c.VolumeMounts {
-				if mount.MountPath == serviceAccountMountPath {
-					foundSATokenMount = true
-					break
-				}
-			}
-			if !foundSATokenMount {
-				t.Fatalf("expected ServiceAccount token mount at %s", serviceAccountMountPath)
-			}
-
-			if c.StartupProbe == nil || c.StartupProbe.Exec == nil {
-				t.Fatalf("expected container to have startup probe")
-			}
-			if c.LivenessProbe == nil || c.LivenessProbe.Exec == nil {
-				t.Fatalf("expected container to have liveness probe")
-			}
-			if c.ReadinessProbe == nil || c.ReadinessProbe.Exec == nil {
-				t.Fatalf("expected container to have readiness probe")
-			}
-
-			if len(c.StartupProbe.Exec.Command) == 0 || c.StartupProbe.Exec.Command[0] != constants.PathProbeBinary {
-				t.Fatalf("expected startup probe to exec %s, got %v", constants.PathProbeBinary, c.StartupProbe.Exec.Command)
-			}
-			if len(c.LivenessProbe.Exec.Command) == 0 || c.LivenessProbe.Exec.Command[0] != constants.PathProbeBinary {
-				t.Fatalf("expected liveness probe to exec %s, got %v", constants.PathProbeBinary, c.LivenessProbe.Exec.Command)
-			}
-			if len(c.ReadinessProbe.Exec.Command) == 0 || c.ReadinessProbe.Exec.Command[0] != constants.PathProbeBinary {
-				t.Fatalf("expected readiness probe to exec %s, got %v", constants.PathProbeBinary, c.ReadinessProbe.Exec.Command)
-			}
-
-			if c.StartupProbe.TimeoutSeconds != 10 {
-				t.Fatalf("expected startup probe timeout to be 10s, got %d", c.StartupProbe.TimeoutSeconds)
-			}
-
-			if !strings.Contains(strings.Join(c.StartupProbe.Exec.Command, " "), "-addr="+testOpenBaoProbeAddr) {
-				t.Fatalf("expected startup probe to target %s, got %v", testOpenBaoProbeAddr, c.StartupProbe.Exec.Command)
-			}
-			if !strings.Contains(strings.Join(c.StartupProbe.Exec.Command, " "), "-mode=startup") {
-				t.Fatalf("expected startup probe to use startup mode, got %v", c.StartupProbe.Exec.Command)
-			}
-			if !strings.Contains(strings.Join(c.StartupProbe.Exec.Command, " "), "-timeout="+testOpenBaoStartupProbeTimeout) {
-				t.Fatalf("expected startup probe to use timeout %s, got %v", testOpenBaoStartupProbeTimeout, c.StartupProbe.Exec.Command)
-			}
-
-			if !strings.Contains(strings.Join(c.LivenessProbe.Exec.Command, " "), "-addr="+testOpenBaoProbeAddr) {
-				t.Fatalf("expected liveness probe to target %s, got %v", testOpenBaoProbeAddr, c.LivenessProbe.Exec.Command)
-			}
-			if !strings.Contains(strings.Join(c.LivenessProbe.Exec.Command, " "), "-timeout="+testOpenBaoLivenessProbeTimeout) {
-				t.Fatalf("expected liveness probe to use timeout %s, got %v", testOpenBaoLivenessProbeTimeout, c.LivenessProbe.Exec.Command)
-			}
-
-			if !strings.Contains(strings.Join(c.ReadinessProbe.Exec.Command, " "), "-addr="+testOpenBaoProbeAddr) {
-				t.Fatalf("expected readiness probe to target %s, got %v", testOpenBaoProbeAddr, c.ReadinessProbe.Exec.Command)
-			}
-			if !strings.Contains(strings.Join(c.ReadinessProbe.Exec.Command, " "), "-timeout="+testOpenBaoReadinessProbeTimeout) {
-				t.Fatalf("expected readiness probe to use timeout %s, got %v", testOpenBaoReadinessProbeTimeout, c.ReadinessProbe.Exec.Command)
-			}
-
-			// Verify environment variables are set correctly
-			envVars := make(map[string]string)
-			for _, env := range c.Env {
-				if env.ValueFrom != nil {
-					if env.ValueFrom.FieldRef != nil {
-						envVars[env.Name] = env.ValueFrom.FieldRef.FieldPath
-					}
-				} else if env.Value != "" {
-					envVars[env.Name] = env.Value
-				}
-			}
-
-			if envVars[constants.EnvHostname] != "metadata.name" {
-				t.Fatalf("expected %s env var to reference metadata.name, got %q", constants.EnvHostname, envVars[constants.EnvHostname])
-			}
-			if envVars[constants.EnvBaoK8sPodName] != "metadata.name" {
-				t.Fatalf("expected %s env var to reference metadata.name, got %q", constants.EnvBaoK8sPodName, envVars[constants.EnvBaoK8sPodName])
-			}
-			if envVars[constants.EnvBaoK8sNamespace] != "metadata.namespace" {
-				t.Fatalf("expected %s env var to reference metadata.namespace, got %q", constants.EnvBaoK8sNamespace, envVars[constants.EnvBaoK8sNamespace])
-			}
-			if envVars[constants.EnvPodIP] != "status.podIP" {
-				t.Fatalf("expected %s env var to reference status.podIP, got %q", constants.EnvPodIP, envVars[constants.EnvPodIP])
-			}
-			if !strings.Contains(envVars[constants.EnvBaoAPIAddr], "$("+constants.EnvPodIP+")") {
-				t.Fatalf("expected %s env var to contain $(%s), got %q", constants.EnvBaoAPIAddr, constants.EnvPodIP, envVars[constants.EnvBaoAPIAddr])
-			}
-			if envVars["UMASK"] != "0077" {
-				t.Fatalf("expected UMASK env var to be 0077, got %q", envVars["UMASK"])
-			}
-		}
-	}
-	if !containerFound {
-		t.Fatalf("expected StatefulSet to have container %q", constants.ContainerBao)
-	}
-}
-
-func TestProbesUseACMEDomainWhenACMEEnabled(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("acme-probe", ns)
-	cluster.Spec.Replicas = 1
-	cluster.Spec.TLS.Mode = openbaov1alpha1.TLSModeACME
-	acmeDomain := fmt.Sprintf("%s.%s.svc", cluster.Name, cluster.Namespace)
-	cluster.Spec.TLS.ACME = &openbaov1alpha1.ACMEConfig{
-		DirectoryURL: "https://example.com/acme",
-		Domains:      []string{acmeDomain},
-		Email:        "e2e@example.invalid",
-	}
-	// Configure tls_acme_ca_root to use the ACME CA for certificate verification
-	cluster.Spec.Configuration = &openbaov1alpha1.OpenBaoConfiguration{
-		ACMECARoot: "/etc/bao/seal-creds/ca.crt",
-	}
-	cluster.Spec.Unseal = &openbaov1alpha1.UnsealConfig{
-		Type: "transit",
-		Transit: &openbaov1alpha1.TransitSealConfig{
-			Address:   "https://openbao:8200",
-			KeyName:   "transit-key",
-			MountPath: "transit/",
-		},
-		CredentialsSecretRef: &corev1.LocalObjectReference{
-			Name: "infra-bao-token",
-		},
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Provide the seal-creds secret so the CA file path exists.
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "infra-bao-token",
-			Namespace: cluster.Namespace,
-		},
-		Data: map[string][]byte{
-			"token":      []byte("root"),
-			"ca.crt":     []byte("dummy"),
-			"pki-ca.crt": newTestCACertPEM(t),
-		},
-	}
-	if err := k8sClient.Create(context.Background(), secret); err != nil {
-		t.Fatalf("failed to create credentials secret: %v", err)
-	}
-
-	ctx := context.Background()
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet); err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	var probesFound bool
-	for _, c := range statefulSet.Spec.Template.Spec.Containers {
-		if c.Name != constants.ContainerBao {
-			continue
-		}
-		probesFound = true
-		cmd := strings.Join(c.ReadinessProbe.Exec.Command, " ")
-		// In ACME mode, probes should still use loopback address but set SNI to the ACME domain
-		if !strings.Contains(cmd, "-addr="+testOpenBaoProbeAddr) {
-			t.Fatalf("expected readiness probe to use loopback address %s, got %v", testOpenBaoProbeAddr, c.ReadinessProbe.Exec.Command)
-		}
-		if !strings.Contains(cmd, "-servername="+acmeDomain) {
-			t.Fatalf("expected readiness probe to set SNI to ACME domain, got %v", c.ReadinessProbe.Exec.Command)
-		}
-		// In ACME mode with tls_acme_ca_root configured, probes should use the PKI CA file derived from it.
-		if !strings.Contains(cmd, "-ca-file=/etc/bao/seal-creds/pki-ca.crt") {
-			t.Fatalf("expected readiness probe to use derived PKI CA file from tls_acme_ca_root, got %v", c.ReadinessProbe.Exec.Command)
-		}
-	}
-	if !probesFound {
-		t.Fatalf("expected to find openbao container probes")
-	}
-}
-
-func TestProbesUseACMEDomainWhenACMEEnabled_PublicACME(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("acme-public", ns)
-	cluster.Spec.Replicas = 1
-	cluster.Spec.TLS.Mode = openbaov1alpha1.TLSModeACME
-	cluster.Spec.TLS.ACME = &openbaov1alpha1.ACMEConfig{
-		DirectoryURL: "https://acme-v02.api.letsencrypt.org/directory",
-		Domain:       "example.com",
-		Email:        "admin@example.com",
-	}
-	// No Unseal config, so no seal-creds volume - this simulates public ACME
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet); err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	var probesFound bool
-	for _, c := range statefulSet.Spec.Template.Spec.Containers {
-		if c.Name != constants.ContainerBao {
-			continue
-		}
-		probesFound = true
-		cmd := strings.Join(c.ReadinessProbe.Exec.Command, " ")
-		// In ACME mode, probes should still use loopback address but set SNI to the ACME domain
-		if !strings.Contains(cmd, "-addr="+testOpenBaoProbeAddr) {
-			t.Fatalf("expected readiness probe to use loopback address %s, got %v", testOpenBaoProbeAddr, c.ReadinessProbe.Exec.Command)
-		}
-		if !strings.Contains(cmd, "-servername=example.com") {
-			t.Fatalf("expected readiness probe to set SNI to ACME domain, got %v", c.ReadinessProbe.Exec.Command)
-		}
-		// For public ACME, no CA file should be specified (uses system roots)
-		if strings.Contains(cmd, "-ca-file=") {
-			t.Fatalf("expected readiness probe to NOT specify CA file for public ACME, got %v", c.ReadinessProbe.Exec.Command)
-		}
-	}
-	if !probesFound {
-		t.Fatalf("expected to find openbao container probes")
-	}
-}
-
-func TestProbesSetSNIToExternalServiceWhenServiceEnabled_NonACME(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("probe-sni", ns)
-	cluster.Status.Initialized = true
-	cluster.Spec.Service = &openbaov1alpha1.ServiceConfig{
-		Type: corev1.ServiceTypeClusterIP,
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	if err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet); err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	wantServerName := "-servername=" + portopenbao.ComputeTLSServerName(cluster)
-
-	var probesFound bool
-	for _, c := range statefulSet.Spec.Template.Spec.Containers {
-		if c.Name != constants.ContainerBao {
-			continue
-		}
-		probesFound = true
-		cmd := strings.Join(c.ReadinessProbe.Exec.Command, " ")
-		if !strings.Contains(cmd, wantServerName) {
-			t.Fatalf("expected readiness probe to set SNI to internal TLS server name (%s), got %v", wantServerName, c.ReadinessProbe.Exec.Command)
-		}
-	}
-	if !probesFound {
-		t.Fatalf("expected to find openbao container probes")
-	}
-}
-
-func TestStatefulSetHasInitContainerWhenEnabled(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-init-container", ns)
-	cluster.Spec.InitContainer = &openbaov1alpha1.InitContainerConfig{
-		Enabled: true,
-		Image:   "openbao/openbao-init:latest",
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if len(statefulSet.Spec.Template.Spec.InitContainers) == 0 {
-		t.Fatalf("expected StatefulSet to have init containers when enabled")
-	}
-
-	initContainerFound := false
-	for _, ic := range statefulSet.Spec.Template.Spec.InitContainers {
-		if ic.Name == "bao-config-init" {
-			initContainerFound = true
-			if ic.Image != "openbao/openbao-init:latest" {
-				t.Fatalf("expected init container image %q, got %q", "openbao/openbao-init:latest", ic.Image)
-			}
-		}
-	}
-	if !initContainerFound {
-		t.Fatalf("expected StatefulSet to have bao-config-init container")
-	}
-}
-
-func TestStatefulSetIncludesInitContainerEvenWhenDisabledFlagSet(t *testing.T) {
-	// Set OPERATOR_VERSION to avoid panic from fail-fast logic in DefaultInitImage
-	t.Setenv(constants.EnvOperatorVersion, "1.0.0")
-
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-no-init-container", ns)
-	cluster.Spec.InitContainer = &openbaov1alpha1.InitContainerConfig{
-		Enabled: false,
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	if len(statefulSet.Spec.Template.Spec.InitContainers) == 0 {
-		t.Fatalf("expected StatefulSet to always include init containers even when InitContainer.Enabled is false")
-	}
-}
-
-func TestStatefulSetHasCorrectVolumeMounts(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-volumes", ns)
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	// Create TLS secret before Reconcile, as ensureStatefulSet now checks for prerequisites
-	createTLSSecretForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.reconcileWithWorkload(ctx, logr.Discard(), cluster, spec); err != nil {
-		t.Fatalf("Reconcile() error = %v", err)
-	}
-
-	statefulSet := &appsv1.StatefulSet{}
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      statefulSetName(cluster),
-	}, statefulSet)
-	if err != nil {
-		t.Fatalf("expected StatefulSet to exist: %v", err)
-	}
-
-	expectedVolumes := []string{
-		tlsVolumeName,
-		configVolumeName,
-		configRenderedVolumeName,
-		unsealVolumeName,
-		dataVolumeName,
-	}
-
-	containerFound := false
-	for _, c := range statefulSet.Spec.Template.Spec.Containers {
-		if c.Name == constants.ContainerBao {
-			containerFound = true
-			volumeMountNames := make(map[string]bool)
-			for _, vm := range c.VolumeMounts {
-				volumeMountNames[vm.Name] = true
-			}
-
-			for _, expectedVol := range expectedVolumes {
-				if !volumeMountNames[expectedVol] {
-					t.Fatalf("expected container to have volume mount %q", expectedVol)
-				}
-			}
-		}
-	}
-	if !containerFound {
-		t.Fatalf("expected StatefulSet to have container %q", constants.ContainerBao)
-	}
-}
-
-func TestDeletePVCsDeletesAllPVCs(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-delete-pvcs", ns)
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	// Create PVCs
-	pvc1 := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "data-infra-delete-pvcs-0",
-			Namespace: cluster.Namespace,
-			Labels: map[string]string{
-				constants.LabelOpenBaoCluster: cluster.Name,
-			},
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceStorage: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-	pvc2 := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "data-infra-delete-pvcs-1",
-			Namespace: cluster.Namespace,
-			Labels: map[string]string{
-				constants.LabelOpenBaoCluster: cluster.Name,
-			},
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceStorage: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-
-	if err := k8sClient.Create(ctx, pvc1); err != nil {
-		t.Fatalf("failed to create PVC1: %v", err)
-	}
-	if err := k8sClient.Create(ctx, pvc2); err != nil {
-		t.Fatalf("failed to create PVC2: %v", err)
-	}
-
-	// Cleanup with DeletePVCs policy
-	if err := manager.Cleanup(ctx, logr.Discard(), cluster, openbaov1alpha1.DeletionPolicyDeletePVCs); err != nil {
-		t.Fatalf("Cleanup() error = %v", err)
-	}
-
-	// Verify PVCs are deleted
-	err := k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      pvc1.Name,
-	}, &corev1.PersistentVolumeClaim{})
-	if !apierrors.IsNotFound(err) {
-		t.Fatalf("expected PVC1 to be deleted, got error: %v", err)
-	}
-
-	err = k8sClient.Get(ctx, types.NamespacedName{
-		Namespace: cluster.Namespace,
-		Name:      pvc2.Name,
-	}, &corev1.PersistentVolumeClaim{})
-	if !apierrors.IsNotFound(err) {
-		t.Fatalf("expected PVC2 to be deleted, got error: %v", err)
-	}
-}
-
-func TestDeletePVCsPreservesExistingACMESharedCachePVC(t *testing.T) {
-	k8sClient, _ := envtestClientForPackage(t)
-	manager := NewManager(k8sClient)
-
-	ns := testNamespace(t)
-	cluster := newMinimalCluster("infra-delete-acme-cache", ns)
-	cluster.Spec.TLS.Mode = openbaov1alpha1.TLSModeACME
-	cluster.Spec.TLS.ACME = &openbaov1alpha1.ACMEConfig{
-		DirectoryURL: "https://example.invalid/acme",
-		Domain:       "example.com",
-		SharedCache: &openbaov1alpha1.ACMESharedCacheConfig{
-			Mode:              openbaov1alpha1.ACMESharedCacheModeExistingPVC,
-			ExistingClaimName: "shared-acme-cache",
-		},
-	}
-	createClusterCRForTest(t, k8sClient, cluster)
-
-	ctx := context.Background()
-
-	dataPVC := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "data-infra-delete-acme-cache-0",
-			Namespace: cluster.Namespace,
-			Labels: map[string]string{
-				constants.LabelOpenBaoCluster: cluster.Name,
-			},
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceStorage: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-	cachePVC := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "shared-acme-cache",
-			Namespace: cluster.Namespace,
-			Labels: map[string]string{
-				constants.LabelOpenBaoCluster: cluster.Name,
-			},
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteMany},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceStorage: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-
-	if err := k8sClient.Create(ctx, dataPVC); err != nil {
-		t.Fatalf("failed to create data PVC: %v", err)
-	}
-	if err := k8sClient.Create(ctx, cachePVC); err != nil {
-		t.Fatalf("failed to create cache PVC: %v", err)
-	}
-
-	if err := manager.Cleanup(ctx, logr.Discard(), cluster, openbaov1alpha1.DeletionPolicyDeletePVCs); err != nil {
-		t.Fatalf("Cleanup() error = %v", err)
-	}
-
-	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: cluster.Namespace, Name: dataPVC.Name}, &corev1.PersistentVolumeClaim{}); !apierrors.IsNotFound(err) {
-		t.Fatalf("expected data PVC to be deleted, got error: %v", err)
-	}
-	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: cluster.Namespace, Name: cachePVC.Name}, &corev1.PersistentVolumeClaim{}); err != nil {
-		t.Fatalf("expected existing ACME shared cache PVC to be preserved, got error: %v", err)
-	}
-}
