--- internal/service/infra/helpers_test.go
@@ -1,111 +1,0 @@
-//go:build integration
-// +build integration
-
-package infra
-
-import (
-	"testing"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
-
-	openbaov1alpha1 "github.com/dc-tec/openbao-operator/api/v1alpha1"
-	"github.com/dc-tec/openbao-operator/internal/platform/constants"
-)
-
-// testScheme is a shared scheme used across tests.
-var testScheme = func() *runtime.Scheme {
-	scheme := runtime.NewScheme()
-	_ = clientgoscheme.AddToScheme(scheme)
-	_ = gatewayv1.Install(scheme)
-	_ = openbaov1alpha1.AddToScheme(scheme)
-	return scheme
-}()
-
-func newTestClient(t *testing.T) client.Client {
-	t.Helper()
-
-	// Create the Kubernetes service that NetworkPolicy detection requires
-	kubernetesService := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kubernetes",
-			Namespace: "default",
-		},
-		Spec: corev1.ServiceSpec{
-			ClusterIP: "10.43.0.1", // Used to derive the kubernetes Service IP CIDR (10.43.0.1/32)
-			Ports: []corev1.ServicePort{
-				{
-					Name: "https",
-					Port: 443,
-				},
-			},
-		},
-	}
-
-	return fake.NewClientBuilder().
-		WithScheme(testScheme).
-		WithObjects(kubernetesService).
-		WithReturnManagedFields().
-		Build()
-}
-
-const (
-	apiVersion               = "openbao.org/v1alpha1"
-	kind                     = "OpenBaoCluster"
-	dataVolumeName           = constants.VolumeData
-	tlsVolumeName            = constants.VolumeTLS
-	configVolumeName         = constants.VolumeConfig
-	configRenderedVolumeName = "config-rendered"
-	unsealVolumeName         = "unseal"
-	serviceAccountMountPath  = "/var/run/secrets/kubernetes.io/serviceaccount"
-	openBaoBinaryName        = constants.BinaryBao
-)
-
-//nolint:unparam // namespace is used in other tests/integration tests
-func newMinimalCluster(name, namespace string) *openbaov1alpha1.OpenBaoCluster {
-	return &openbaov1alpha1.OpenBaoCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: openbaov1alpha1.OpenBaoClusterSpec{
-			Version:  "2.4.4",
-			Image:    "openbao/openbao:2.4.4",
-			Replicas: 3,
-			TLS: openbaov1alpha1.TLSConfig{
-				Enabled:        true,
-				RotationPeriod: "720h",
-			},
-			Storage: openbaov1alpha1.StorageConfig{
-				Size: "10Gi",
-			},
-			InitContainer: &openbaov1alpha1.InitContainerConfig{
-				Image: "openbao/openbao-init:latest",
-			},
-		},
-	}
-}
-
-func unsealSecretName(cluster *openbaov1alpha1.OpenBaoCluster) string {
-	return cluster.Name + constants.SuffixUnsealKey
-}
-
-func configMapName(cluster *openbaov1alpha1.OpenBaoCluster) string {
-	return cluster.Name + constants.SuffixConfigMap
-}
-
-func headlessServiceName(cluster *openbaov1alpha1.OpenBaoCluster) string {
-	return cluster.Name
-}
-
-func serviceAccountName(cluster *openbaov1alpha1.OpenBaoCluster) string {
-	if cluster.Spec.ServiceAccount != nil && cluster.Spec.ServiceAccount.Name != "" {
-		return cluster.Spec.ServiceAccount.Name
-	}
-	return cluster.Name + constants.SuffixServiceAccount
-}
