+++ test/integration/infra_full_reconcile_test.go
@@ -27,10 +27,9 @@ func TestInfraFullReconcile_StatefulSet_SSAAndIdempotency(t *testing.T) {
 	})
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -88,7 +87,7 @@ func TestInfraFullReconcile_StatefulSet_SSAAndIdempotency(t *testing.T) {
 	}
 
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after drift error = %v", err)
 	}
 
@@ -124,7 +123,7 @@ func TestInfraFullReconcile_StatefulSet_SSAAndIdempotency(t *testing.T) {
 	}
 
 	// One more pass should be safe/idempotent.
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() third call error = %v", err)
 	}
 }
