+++ test/integration/infra_network_test.go
@@ -38,13 +38,12 @@ func TestInfraNetwork_HeadlessService_IsIdempotent(t *testing.T) {
 	createTLSSecret(t, namespace, cluster.Name)
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() second call error = %v", err)
 	}
 
@@ -79,9 +78,8 @@ func TestInfraNetwork_ExternalService_CreatesAndDeletes(t *testing.T) {
 	createTLSSecret(t, namespace, cluster.Name)
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -105,7 +103,7 @@ func TestInfraNetwork_ExternalService_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling external access error = %v", err)
 	}
 
@@ -130,9 +128,8 @@ func TestInfraNetwork_Ingress_CreatesAndDeletes(t *testing.T) {
 	createTLSSecret(t, namespace, cluster.Name)
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -147,7 +144,7 @@ func TestInfraNetwork_Ingress_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling ingress error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: cluster.Name}, ing); err == nil {
@@ -174,9 +171,8 @@ func TestInfraNetwork_HTTPRoute_CreatesAndDeletes(t *testing.T) {
 	createTLSSecret(t, namespace, cluster.Name)
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -198,7 +194,7 @@ func TestInfraNetwork_HTTPRoute_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling gateway error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: routeName}, httpRoute); err == nil {
@@ -228,9 +224,8 @@ func TestInfraNetwork_GatewayCAConfigMap_CreatesUpdatesAndDeletes(t *testing.T) 
 	createCASecret(t, namespace, cluster.Name, ca1)
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -255,7 +250,7 @@ func TestInfraNetwork_GatewayCAConfigMap_CreatesUpdatesAndDeletes(t *testing.T) 
 		t.Fatalf("update CA secret: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after CA update error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: cmName}, cm); err != nil {
@@ -271,7 +266,7 @@ func TestInfraNetwork_GatewayCAConfigMap_CreatesUpdatesAndDeletes(t *testing.T) 
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, logr.Discard(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, logr.Discard(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling gateway error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: cmName}, cm); err == nil {
@@ -339,9 +334,8 @@ func TestInfraNetwork_BlueGreenExternalService_UsesRevisionSelectorAndCleansStal
 	})
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -380,7 +374,7 @@ func TestInfraNetwork_BlueGreenExternalService_UsesRevisionSelectorAndCleansStal
 		status.BlueGreen.Phase = openbaov1alpha1.PhaseDemotingBlue
 	})
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() during demoting blue error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: cluster.Name + infraPublicServiceSuffix}, mainSvc); err != nil {
@@ -394,7 +388,7 @@ func TestInfraNetwork_BlueGreenExternalService_UsesRevisionSelectorAndCleansStal
 		status.BlueGreen.Phase = openbaov1alpha1.PhaseCleanup
 	})
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() during cleanup error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: cluster.Name + infraPublicServiceSuffix}, mainSvc); err != nil {
@@ -425,9 +419,8 @@ func TestInfraNetwork_TLSRoute_CreatesAndDeletes(t *testing.T) {
 	createCASecret(t, namespace, cluster.Name, []byte("ca-1"))
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -474,7 +467,7 @@ func TestInfraNetwork_TLSRoute_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling TLS passthrough error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: tlsRouteName}, tlsRoute); err == nil {
@@ -505,9 +498,8 @@ func TestInfraNetwork_BackendTLSPolicy_CreatesAndDeletes(t *testing.T) {
 	createCASecret(t, namespace, cluster.Name, []byte("ca-1"))
 
 	controllerClient := newControllerClient(t)
-	manager := newIntegrationInfraManager(controllerClient, k8sScheme)
 	spec := newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() error = %v", err)
 	}
 
@@ -534,7 +526,7 @@ func TestInfraNetwork_BackendTLSPolicy_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after disabling BackendTLS error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: backendTLSName}, backendTLS); err == nil {
@@ -550,7 +542,7 @@ func TestInfraNetwork_BackendTLSPolicy_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after enabling BackendTLS error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: backendTLSName}, backendTLS); err != nil {
@@ -563,7 +555,7 @@ func TestInfraNetwork_BackendTLSPolicy_CreatesAndDeletes(t *testing.T) {
 		t.Fatalf("update cluster: %v", err)
 	}
 	spec = newTestStatefulSetSpec(cluster)
-	if err := manager.Reconcile(ctx, discardLogger(), cluster, spec); err != nil {
+	if err := reconcileClusterResources(ctx, discardLogger(), controllerClient, k8sScheme, cluster, spec); err != nil {
 		t.Fatalf("Reconcile() after enabling TLS passthrough error = %v", err)
 	}
 	if err := k8sClient.Get(ctx, types.NamespacedName{Namespace: namespace, Name: backendTLSName}, backendTLS); err == nil {
